Feat: Chat width split manage dynamically Feat: Split stream screen in small files for clean
